+++ src/app/orders/orders.module.ts
@@ -4,16 +4,21 @@ import { CommonModule } from '@angular/common';
 import { OrdersRoutingModule } from './orders-routing.module';
 import { ordersAPIProvider } from './orders.config';
 import { SharedModule } from '../shared/shared.module';
+import { ProcessOrderComponent } from './components/process-order/process-order.component';
+import { ReactiveFormsModule } from '@angular/forms';
+import { processOrderConfigProvider } from './process-order-config';
 
 @NgModule({
-  declarations: [OrdersRoutingModule.components],
+  declarations: [OrdersRoutingModule.components, ProcessOrderComponent],
   imports: [
     CommonModule,
     SharedModule,
+    ReactiveFormsModule,
     OrdersRoutingModule
   ],
   providers: [
-    ordersAPIProvider
+    ordersAPIProvider,
+    processOrderConfigProvider
   ]
 })
 export class OrdersModule {